refactor: extract gallery events from AgentWorkspace
