+++ src/components/agent/use-agent-gallery-events-types.ts
@@ -1,0 +1,7 @@
+import { type AgentChatGallerySection } from "../../lib/agent-chat-gallery";
+import type * as React from "react";
+
+export type UseAgentGalleryEventsDependencies = {
+  setGalleryErrors: React.Dispatch<React.SetStateAction<boolean>>;
+  setGallerySections: React.Dispatch<React.SetStateAction<AgentChatGallerySection[] | null>>;
+};
